GPUパーティクルシステムの多機能化・物理拡張
パーティクルエミッターにRing/Cylinder型を追加し、寿命・スケール・カラー・回転などのランダマイズや物理挙動（重力・空気抵抗・床衝突・アトラクター）を拡張。パーティクルごとにカラー・スケール・回転を補間し、ビルボード/3Dメッシュ形状・アトラスアニメーションにも対応。ImGuiデバッグUIで全パラメータをリアルタイム調整可能にし、C++/HLSL両側の構造体・APIを整理。DrawManagerも任意メッシュのインスタンシング描画に対応。全体的に現代的なVFX表現に必要な機能を網羅。

+++ project/Application/resources/shaders/InitializeParticle.CS.hlsl
@@ -1,18 +1,15 @@
-
 #include "ParticleGPU.hlsli"
 
-static const uint32_t kMaxParticles = 32768;
+static const uint kMaxParticles = 32768;
 
 RWStructuredBuffer<Particle> gParticles : register(u0);
-
-RWStructuredBuffer<int32_t> gFreeListIndex : register(u1);
-
-RWStructuredBuffer<int32_t> gFreeList : register(u2);
+RWStructuredBuffer<int> gFreeListIndex : register(u1);
+RWStructuredBuffer<int> gFreeList : register(u2);
 
 [numthreads(1024, 1, 1)]
 void main(uint3 DTid : SV_DispatchThreadID)
 {
-	uint32_t particleIndex = DTid.x;
+	uint particleIndex = DTid.x;
 	
 	if (particleIndex == 0)
 	{
@@ -21,9 +18,8 @@ void main(uint3 DTid : SV_DispatchThreadID)
 	
 	if (particleIndex < kMaxParticles)
 	{
-		// Particle構造体の全要素を0で埋めるという書き方
+		// Particle構造体の全要素を0で埋める
 		gParticles[particleIndex] = (Particle) 0;
-		gFreeList[particleIndex] = particleIndex;
-
+		gFreeList[particleIndex] = (int)particleIndex;
 	}
 }
